+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/jss/compare/v1.12.1...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Implements `workbox` to handle service worker
 - Implements `webpack` for bundling static files and library
 
 ### Fixed